Rearranged code. Move code to marc and export packages
