+++ .gitignore
@@ -1,2 +1,3 @@
 /tmp/*
 /brown/*
+.DS_Store
