+++ marc/leader.go
@@ -1,4 +1,4 @@
-package main
+package marc
 
 import (
 	"errors"
